requirements.txt
fastapi
uvicorn
openai==0.28.1
faiss-cpu
numpy
requests
PyMuPDF
python-multipart
pydantic
